+++ packages/giu/src/components/backdrop.js
@@ -5,8 +5,6 @@ import { merge } from 'timm';
 import { cancelEvent } from '../gral/helpers';
 import { IS_IOS } from '../gral/constants';
 
-require('./backdrop.css');
-
 // ==========================================
 // Component
 // ==========================================
@@ -19,14 +17,17 @@ class Backdrop extends React.PureComponent {
   props: Props;
 
   render() {
+    // Due to bug: https://github.com/zeit/styled-jsx/issues/329
     return (
       <div
-        className="giu-backdrop"
         onWheel={cancelEvent}
         onTouchMove={cancelEvent}
         {...this.props}
+        className="giu-backdrop"
         style={style.backdrop(this.props)}
-      />
+      >
+        {STYLES}
+      </div>
     );
   }
 }
@@ -48,6 +49,25 @@ const style = {
       baseStyle
     ),
 };
+
+const STYLES = (
+  <style jsx global>{`
+    @keyframes appear {
+      from {
+        opacity: 0;
+      }
+      to {
+        opacity: 0.6;
+      }
+    }
+
+    .giu-backdrop {
+      animation: appear 300ms;
+      -webkit-animation: appear 300ms;
+      opacity: 0.6;
+    }
+  `}</style>
+);
 
 // ==========================================
 // Public API