+++ packages/giu/src/components/modals.js
@@ -10,6 +10,7 @@ import { MISC } from '../gral/constants';
 import type { Action } from '../gral/types';
 import Modal from './modal';
 import type { ModalPars } from './modalTypes';
+import { GLOBAL_CSS } from '../gral/resetCss';
 
 /* --
 **Include the `<Modals />` component at (or near)
@@ -147,6 +148,9 @@ class Modals extends React.PureComponent {
             {...props}
           />
         ))}
+        <style jsx global>
+          {GLOBAL_CSS}
+        </style>
       </div>
     );
   }