+++ packages/giu/src/gral/resetCss.js
@@ -1,0 +1,11 @@
+import css from 'styled-jsx/css'; // eslint-disable-line
+
+export const GLOBAL_CSS = css`
+  *,
+  *:before,
+  *:after {
+    -moz-box-sizing: border-box;
+    -webkit-box-sizing: border-box;
+    box-sizing: border-box;
+  }
+`;